Removing legacy Fortran statements.

+++ trunk/src/solver/pardiso/sparse_storage_frmt.F90
@@ -155,9 +155,9 @@ subroutine assemble_triplet(row,col,val,nz,nnz)
 !..declare variables
    integer    , intent(inout) :: row(nz),col(nz)
 #if C_MODE
-   complex*16 , intent(inout) :: val(nz)
+   complex(8) , intent(inout) :: val(nz)
 #else
-   real*8     , intent(inout) :: val(nz)
+   real(8)    , intent(inout) :: val(nz)
 #endif
    integer    , intent(in)    :: nz
    integer    , intent(out)   :: nnz
@@ -227,11 +227,11 @@ recursive subroutine qsort_triplet_old(ia,ja,xa,n,first,last)
    integer     , intent(in)    :: n,first,last
    integer     , intent(inout) :: ia(n), ja(n)
 #if C_MODE   
-   complex*16  , intent(inout) :: xa(n)
-   complex*16                  :: x
+   complex(8)  , intent(inout) :: xa(n)
+   complex(8)                  :: x
 #else
-   real*8      , intent(inout) :: xa(n)
-   real*8                      :: x
+   real(8)     , intent(inout) :: xa(n)
+   real(8)                     :: x
 #endif
    integer                     :: i,j,k,l
 !
@@ -289,11 +289,11 @@ recursive subroutine qsort_triplet(ia,ja,xa,n,first,last)
    integer     , intent(in)    :: n,first,last
    integer     , intent(inout) :: ia(n), ja(n)
 #if C_MODE   
-   complex*16  , intent(inout) :: xa(n)
-   complex*16                  :: x
+   complex(8)  , intent(inout) :: xa(n)
+   complex(8)                  :: x
 #else
-   real*8      , intent(inout) :: xa(n)
-   real*8                      :: x
+   real(8)     , intent(inout) :: xa(n)
+   real(8)                     :: x
 #endif
    integer                     :: i,j,ki,kj,l
 !
@@ -354,11 +354,11 @@ recursive subroutine partition_triplet_omp(ip,jp,n,k,ia,ja,xa)
    integer    , intent(out)    :: k
    integer    , intent(inout)  :: ia(n), ja(n)
 #if C_MODE   
-   complex*16 , intent(inout)  :: xa(n)
-   complex*16                  :: xaux
+   complex(8) , intent(inout)  :: xa(n)
+   complex(8)                  :: xaux
 #else
-   real*8     , intent(inout)  :: xa(n)
-   real*8                      :: xaux
+   real(8)    , intent(inout)  :: xa(n)
+   real(8)                     :: xaux
 #endif
    integer                     :: i,ig,iaux,jaux,kl,kr,kto
 !..tuning parameter
@@ -435,9 +435,9 @@ recursive subroutine qsort_triplet_omp(n,ia,ja,xa)
    integer    , intent(in)    :: n
    integer    , intent(inout) :: ia(n), ja(n)
 #if C_MODE
-   complex*16 , intent(inout) :: xa(n)
+   complex(8) , intent(inout) :: xa(n)
 #else
-   real*8     , intent(inout) :: xa(n)
+   real(8)    , intent(inout) :: xa(n)
 #endif
    integer                    :: i,ig,ip,jp,n_le
    real                       :: r
@@ -496,9 +496,9 @@ subroutine assemble_double(row,val,nz,nnz)
 
    integer       :: row(nz)
 #if C_MODE
-   complex*16    :: Val(nz)
+   complex(8)    :: Val(nz)
 #else
-   real*8        :: Val(nz)            
+   real(8)       :: Val(nz)
 #endif
    integer       :: nz, nnz
    integer       :: first, last, k, i 
@@ -553,9 +553,9 @@ recursive subroutine qsort_double(ia,val,n,first, last)
    integer     :: i, j, n, k, l
    integer     :: ia(n) 
 #if C_MODE   
-   complex*16  :: val(n), x
+   complex(8)  :: val(n), x
 #else
-   real*8      :: val(n), x
+   real(8)     :: val(n), x
 #endif   
    integer :: first, last
 !
